refactor(app): decompose GenericTable details and sorting (#346)
